Put TweeStyler class in its own module
TweeStyler depends on wxWidgets, while the tweelexer module is used by
the command line "twee" tool as well, which shouldn't depend on any
GUI toolkit.
